Minimize the number of API queries
Because the free plan of the API is rate-limited, I had to find a way to
minimize the number of queries which is mostly done because of fetching
token names (trending tokens list can change rapidly thus it has to
fetch new token names).

Fetch the whole token list only once, then filter for the needed tokens
to get it's name.